+++ core/index.php
